requirements.txt
flask
python-dotenv
binance-futures-connector
